The beginnings of proper atom-feed provision from Vanilla.
Originally I had planned that all the implementation would be within the dynasnip, but a few issues gave me second thoughts about this, particularly with regards to configuration. I'm not 100% about how to separate out the configuration from the feed code nicely. Keeping the attributes in the same snip breaks upgrading via direct replacement of the system snips; keeping the attributes in a parallel snip has no precedent and isn't very obvious.

Generating an atom feed is also similar in many ways to rendering a collection of snips. I might investigate this further, with a view to having something along the lines of a "meta renderer" that handles rendering sets of snips similar to current renderers dealing with single snips. We'll see.

For the moment, the pristine app contains a "feed" snip which the user can customise at will, since it uses code within Vanilla itself.