Allow feeds to be restricted to a particular subset of snips.
This sidesteps a deeper issue - not all snips can be rendered directly. For example, attempting to render the `link_to` snip results in an error, and other snips seem to result in infinite loops. This will be addressed later.

Upgrading to the latest `soup` gem gives a controlled mechanism for stubbing the contents of an app soup, which is useful in this case, although I did spend a lot of time kicking around until I figured out that I also needed to explicitly add the contents of the snips I wanted to test in the stubbed soup.