Don't print warnings in app tests
I haven't figured out a good way to avoid warnings about methods in
dynasnips being redefined; this is due to the way the code in ruby snips
is evaluated, redefining the class and its methods every time it's
rendered.

I probably _could_ somehow capture a mapping between the snip name and
the class, and then unload any classes when it comes to re-rendering the
snip, but that seems overkill for now.

This doesn't quite work, because the default rake task doesn't seem to
"un-set" warnings between the two dependent test tasks.