Add a proper 404 mechanism
This addresses errors I was seeing which turned out to be driven by
requests to `/favicon.ico`. These don't match a snip, but that flow was
still handled by the normal layout and `current_snip` snips. However,
the layout expected to be able to generate the page title, but there was
no snip present. I figured it was overall cleaner to have a custom 404
layout that could be used, which doesn't need to all any dynasnips, and
so we don't need the dynasnips to handle the case where
`app.request.snip` is empty anymore.